+++ web/src/components/molecules/Common/PluginUpload/SettingArea.tsx
