+++ web/src/components/organisms/Common/PluginUpload/hooks.ts
@@ -1,7 +1,7 @@
 import { useCallback, useMemo } from "react";
 
 import Message from "@marketplace/components/atoms/Message";
-import type { FileUploadType } from "@marketplace/components/molecules/Common/PluginEditing/PackageArea";
+import type { FileUploadType } from "@marketplace/components/molecules/Common/PluginUpload/PackageArea";
 import {
   useCreatePluginMutation,
   useParsePluginMutation,
