+++ web/src/components/organisms/UpdatePlugin/index.tsx
@@ -1,9 +1,9 @@
 import { UploadRequestOption } from "rc-upload/lib/interface";
 import { useState } from "react";
 
-import type { FileUploadType } from "@marketplace/components/molecules/Common/PluginEditing/PackageArea";
+import type { FileUploadType } from "@marketplace/components/molecules/Common/PluginUpload/PackageArea";
 import UpdatePluginPage from "@marketplace/components/molecules/UpdatePluginPage";
-import useHooks from "@marketplace/components/organisms/Common/PluginEditing/hooks";
+import useHooks from "@marketplace/components/organisms/Common/PluginUpload/hooks";
 
 export type Props = {
   pluginId?: string;
